api/requirements.txt
# This file is autogenerated by pip-compile with Python 3.12
#
amqp==5.2.0
    # via
    #   -r requirements.in
    #   kombu
anyio==4.4.0
    # via httpx
asgiref==3.8.1
    # via django
attrs==24.2.0
    # via
    #   jsonschema
    #   referencing
billiard==4.2.0
    # via celery
celery==5.4.0
    # via -r requirements.in
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
coreapi==2.3.3
    # via -r requirements.in
coreschema==0.0.4
    # via coreapi
django==5.1.1
    # via
    #   -r requirements.in
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-spectacular
djangorestframework==3.15.2
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
    #   drf-spectacular
djangorestframework-simplejwt==5.3.1
    # via -r requirements.in
drf-spectacular==0.27.2
    # via -r requirements.in
h11==0.14.0
    # via httpcore
httpcore==1.0.5
    # via httpx
httpx==0.27.2
    # via -r requirements.in
idna==3.9
    # via
    #   anyio
    #   httpx
    #   requests
inflection==0.5.1
    # via drf-spectacular
itypes==1.2.0
    # via coreapi
jinja2==3.1.4
    # via coreschema
jsonschema==4.23.0
    # via drf-spectacular
jsonschema-specifications==2023.12.1
    # via jsonschema
kombu==5.4.1
    # via celery
markupsafe==2.1.5
    # via jinja2
prompt-toolkit==3.0.47
    # via click-repl
psycopg2==2.9.9
    # via -r requirements.in
pyjwt==2.9.0
    # via djangorestframework-simplejwt
python-dateutil==2.9.0.post0
    # via celery
python-dotenv==1.0.1
    # via -r requirements.in
pyyaml==6.0.2
    # via drf-spectacular
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via coreapi
rpds-py==0.20.0
    # via
    #   jsonschema
    #   referencing
six==1.16.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
sqlparse==0.5.1
    # via django
tzdata==2024.1
    # via celery
uritemplate==4.1.1
    # via
    #   coreapi
    #   drf-spectacular
urllib3==2.2.3
    # via requests
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via prompt-toolkit
